+++ instrumentation/opentelemetry-instrumentation-django/src/opentelemetry/instrumentation/django/middleware/sqlcommenter_middleware.py
@@ -22,10 +22,8 @@ from django import conf, get_version
 from django.db import connections
 from django.db.backends.utils import CursorDebugWrapper
 
-from opentelemetry.instrumentation.utils import (
-    _add_sql_comment,
-    _get_opentelemetry_values,
-)
+from opentelemetry.instrumentation.sqlcommenter_utils import _add_sql_comment
+from opentelemetry.instrumentation.utils import _get_opentelemetry_values
 from opentelemetry.trace.propagation.tracecontext import (
     TraceContextTextMapPropagator,
 )