+++ instrumentation/opentelemetry-instrumentation-flask/src/opentelemetry/instrumentation/flask/__init__.py
@@ -24,6 +24,53 @@ supports Flask-specific features such as:
 * The ``http.route`` Span attribute is set so that one can see which URL rule
   matched a request.
 
+SQLCOMMENTER
+*****************************************
+You can optionally configure Flask instrumentation to enable sqlcommenter which enriches
+the query with contextual information.
+
+Usage
+-----
+
+.. code:: python
+
+    from opentelemetry.instrumentation.flask import FlaskInstrumentor
+
+    FlaskInstrumentor().instrument(enable_commenter=True, commenter_options={})
+
+
+For example,
+::
+
+   FlaskInstrumentor when used with SQLAlchemyInstrumentor or Psycopg2Instrumentor, invoking cursor.execute("select * from auth_users")
+   will lead to sql query "select * from auth_users" but when SQLCommenter is enabled
+   the query will get appended with some configurable tags like "select * from auth_users /*metrics=value*/;"
+
+    Inorder for the commenter to append flask related tags to sql queries, the commenter needs to enabled on
+    the respective  SQLAlchemyInstrumentor or Psycopg2Instrumentor framework too.
+
+SQLCommenter Configurations
+***************************
+We can configure the tags to be appended to the sqlquery log by adding configuration inside commenter_options(default:{}) keyword
+
+framework = True(Default) or False
+
+For example,
+::
+Enabling this flag will add flask and it's version which is /*flask%%3A2.9.3*/
+
+route = True(Default) or False
+
+For example,
+::
+Enabling this flag will add route uri /*route='/home'*/
+
+controller = True(Default) or False
+
+For example,
+::
+Enabling this flag will add controller name /*controller='home_view'*/
+
 Usage
 -----
 
@@ -255,6 +302,8 @@ def _wrapped_before_request(
     request_hook=None,
     tracer=None,
     excluded_urls=None,
+    enable_commenter=True,
+    commenter_options=None,
 ):
     def _before_request():
         if excluded_urls and excluded_urls.url_disabled(flask.request.url):
@@ -300,6 +349,30 @@ def _wrapped_before_request(
         flask_request_environ[_ENVIRON_SPAN_KEY] = span
         flask_request_environ[_ENVIRON_TOKEN] = token
 
+        if enable_commenter:
+            current_context = context.get_current()
+            flask_info = {}
+
+            # https://flask.palletsprojects.com/en/1.1.x/api/#flask.has_request_context
+            if flask and flask.request:
+                if commenter_options.get("framework", True):
+                    flask_info["framework"] = f"flask:{flask.__version__}"
+                if (
+                    commenter_options.get("controller", True)
+                    and flask.request.endpoint
+                ):
+                    flask_info["controller"] = flask.request.endpoint
+                if (
+                    commenter_options.get("route", True)
+                    and flask.request.url_rule
+                    and flask.request.url_rule.rule
+                ):
+                    flask_info["route"] = flask.request.url_rule.rule
+            sqlcommenter_context = context.set_value(
+                "SQLCOMMENTER_ORM_TAGS_AND_VALUES", flask_info, current_context
+            )
+            context.attach(sqlcommenter_context)
+
     return _before_request
 
 
@@ -336,6 +409,8 @@ class _InstrumentedFlask(flask.Flask):
     _tracer_provider = None
     _request_hook = None
     _response_hook = None
+    _enable_commenter = True
+    _commenter_options = None
     _meter_provider = None
 
     def __init__(self, *args, **kwargs):
@@ -374,6 +449,8 @@ class _InstrumentedFlask(flask.Flask):
             _InstrumentedFlask._request_hook,
             tracer,
             excluded_urls=_InstrumentedFlask._excluded_urls,
+            enable_commenter=_InstrumentedFlask._enable_commenter,
+            commenter_options=_InstrumentedFlask._commenter_options,
         )
         self._before_request = _before_request
         self.before_request(_before_request)
@@ -410,6 +487,11 @@ class FlaskInstrumentor(BaseInstrumentor):
             if excluded_urls is None
             else parse_excluded_urls(excluded_urls)
         )
+        enable_commenter = kwargs.get("enable_commenter", True)
+        _InstrumentedFlask._enable_commenter = enable_commenter
+
+        commenter_options = kwargs.get("commenter_options", {})
+        _InstrumentedFlask._commenter_options = commenter_options
         meter_provider = kwargs.get("meter_provider")
         _InstrumentedFlask._meter_provider = meter_provider
         flask.Flask = _InstrumentedFlask
@@ -424,6 +506,8 @@ class FlaskInstrumentor(BaseInstrumentor):
         response_hook=None,
         tracer_provider=None,
         excluded_urls=None,
+        enable_commenter=True,
+        commenter_options=None,
         meter_provider=None,
     ):
         if not hasattr(app, "_is_instrumented_by_opentelemetry"):
@@ -462,6 +546,10 @@ class FlaskInstrumentor(BaseInstrumentor):
                 request_hook,
                 tracer,
                 excluded_urls=excluded_urls,
+                enable_commenter=enable_commenter,
+                commenter_options=commenter_options
+                if commenter_options
+                else {},
             )
             app._before_request = _before_request
             app.before_request(_before_request)