+++ instrumentation/opentelemetry-instrumentation-sqlalchemy/src/opentelemetry/instrumentation/sqlalchemy/engine.py
@@ -20,10 +20,8 @@ from opentelemetry.instrumentation.sqlalchemy.package import (
     _instrumenting_module_name,
 )
 from opentelemetry.instrumentation.sqlalchemy.version import __version__
-from opentelemetry.instrumentation.utils import (
-    _add_sql_comment,
-    _get_opentelemetry_values,
-)
+from opentelemetry.instrumentation.sqlcommenter_utils import _add_sql_comment
+from opentelemetry.instrumentation.utils import _get_opentelemetry_values
 from opentelemetry.semconv.trace import NetTransportValues, SpanAttributes
 from opentelemetry.trace.status import Status, StatusCode