+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/utils.py
@@ -124,29 +124,6 @@ def _start_internal_or_server_span(
     return span, token
 
 
-def _generate_sql_comment(**meta) -> str:
-    """
-    Return a SQL comment with comma delimited key=value pairs created from
-    **meta kwargs.
-    """
-    key_value_delimiter = ","
-
-    if not meta:  # No entries added.
-        return ""
-
-    # Sort the keywords to ensure that caching works and that testing is
-    # deterministic. It eases visual inspection as well.
-    return (
-        " /*"
-        + key_value_delimiter.join(
-            f"{_url_quote(key)}={_url_quote(value)!r}"
-            for key, value in sorted(meta.items())
-            if value is not None
-        )
-        + "*/"
-    )
-
-
 def _url_quote(s) -> str:  # pylint: disable=invalid-name
     if not isinstance(s, (str, bytes)):
         return s
@@ -175,16 +152,3 @@ def _python_path_without_directory(python_path, directory, path_separator):
         "",
         python_path,
     )
-
-
-def _add_sql_comment(sql, **meta) -> str:
-    """
-    Appends comments to the sql statement and returns it
-    """
-    comment = _generate_sql_comment(**meta)
-    sql = sql.rstrip()
-    if sql[-1] == ";":
-        sql = sql[:-1] + comment + ";"
-    else:
-        sql = sql + comment
-    return sql