+++ opentelemetry-instrumentation/tests/test_utils.py
@@ -15,8 +15,8 @@
 import unittest
 from http import HTTPStatus
 
+from opentelemetry.instrumentation.sqlcommenter_utils import _add_sql_comment
 from opentelemetry.instrumentation.utils import (
-    _add_sql_comment,
     _python_path_without_directory,
     http_status_to_status_code,
 )